Conversion_Workflow - Human-in-loop/pyproject.toml
[project]
name = "qmigrator-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.12",
    "langchain-anthropic>=0.1.1",
    "langchain-groq>=0.1.2",
    "langchain-community>=0.2.10",
    "langgraph>=0.3.30",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.0.1",
    "langchain-google-genai>=2.1.3"
]
